seperate search and aggregation logic from work_handler
